refactor: add computer repository save method and add save method in laboratory repository
