+++ Repositories/ComputerRepository.cs
@@ -9,8 +9,7 @@ class ComputerRepository
     private DatabaseConfig databaseConfig;
 
     public ComputerRepository(DatabaseConfig databaseConfig) => this.databaseConfig = databaseConfig;
-  
-
+    
     public List<Computer> GetAll()
     {
         var Computers = new List<Computer>();
@@ -37,5 +36,21 @@ class ComputerRepository
         connection.Close();
 
         return Computers;
+    }
+
+    public void Save(Computer computer)
+    {
+        var connection = new SqliteConnection(databaseConfig.ConnectionString);
+        connection.Open();
+
+        var command = connection.CreateCommand();
+
+        command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor);";
+        command.Parameters.AddWithValue("$id", computer.Id);
+        command.Parameters.AddWithValue("$ram", computer.Ram);
+        command.Parameters.AddWithValue("$processor", computer.Processor);
+        
+        command.ExecuteNonQuery();
+        connection.Close();
     }
 }
